Add javadocs to new classes and refactor usertask method access modifiers
