--- src/main/java/duke/command/SystemCommand.java
@@ -1,7 +1,0 @@
-package duke.command;
-
-public abstract class SystemCommand extends Command {
-    protected static SystemCommand parse(CommandNames command) {
-        return new ByeCommand();
-    }
-}
